Revamp hash computation and message similarity checks.
- Don't include full Content-Type headers in hash computation.
- Don't canonicalise in place any more.
- Ignore headers when checking message similarity.

Justification:

Apparently mailing list servers actually munge Content-Type e.g. by
stripping the quotes from charset="us-ascii".  Section 5.1 of RFC2045
says that either form is valid (and they are equivalent).
Additionally, they can choose a different multipart/mixed boundary
delimiter per recipient, so we want to ignore this delimiter when
computing the message digest hash for duplicate detection.  The easy
solution to both of these is to string everything in the Content-Type
after the semi-colon.  However, we still need the boundary delimiter
for retrieving the body as a string.  This means we cannot
canonicalise the message in place (which was kind of ugly anyway).  So
instead, we copy only the headers we want to use for computing the
digest, into a separate string.

Additionally, we no longer include headers when checking message
similarity, since we already know that headers can vary significantly
between duplicates, and we've already carefully cherry-picked the
headers which should affect duplicate detection.  This allows us to
tighten the default thresholds for size and diff comparisons, which
should make the risk of false positives even tinier.