A catch-all exception handler around the main file processing loop. This improves the situation detailed in issue #39 in so far as that exceptions no longer lead to the abortion of the entire processing batch. Exceptions are output to stderr with stack trace and the filenames being processed, so that the underlying reason and the potentially faulty files can be investigated.
The execution of this is not pretty.
Ideally, exception handlers should be limited to the parts of the code which actually throw exceptions, and exceptions should be handled by name.
This was a quick job to stop the software from needlessly aborting the operation.

The patch was written against the current stable from pip, but applied cleanly to the current development version.

----
Amended to change the formatting of the log message, because Travis CI is stuck in the 90s (80 char width limit).
I have not actually tested this change. I'm sure Travis will complain anew if the syntax is off.