Add Message-ID as a header to check
Even though Message-ID can be non-unique or even non-existent and
thus cannot be used alone to reliably deduplicate, including it
in the canonical header should be harmless. This is especially
beneficial at the moment when time is stripped from header 'Date'.
For example, if there are back-and-forth replies to the same email
subject on the same day, the current canonical header will be
"Date: YYYY/MM/DD UTC\nFrom: Sender <userA@domainA>\nTo: Recipient
<userB@domainB>\nSubject: Re: XYZ...".
DuplicateSet.check_differences() will reject the entire set, even
when there are duplicates in the set.