fix dynamic mailbox inheritance
The DedupMail mixin class went to great lengths to find the proper
superclass to call a constructor on. However, this ended up calling the
ctor of Message for MaildirMessage because super(k, o) doesn't call the
k constructor in the MRO of o, but starts with the next one.

This just fixes the MRO search so that MaildirMessage is the ctor to
call, assuming the dynamic base-class lookup was a good idea to begin
with :)